Extract MatchData specs from Regexp test
